requirements.txt
accelerate==1.6.0
compressai==1.2.6
diffusers==0.33.1
einops==0.8.1
loguru==0.7.3
matplotlib==3.10.0
numpy==1.26.4
omegaconf==2.3.0
open_clip_torch==2.32.0
opencv_python==4.11.0.86
pillow==11.2.1
protobuf==5.29.4
pycryptodome==3.22.0
pytorch_lightning==2.5.1
requests==2.32.3
scikit_image==0.25.2
scikit_learn==1.6.1
sentencepiece==0.2.0
tabulate==0.9.0
tensorboard==2.19.0
torch_fidelity==0.3.0
torchinfo==1.8.0
tqdm==4.67.1
torchaudio==2.7.0
torchvision==0.22.0
transformers==4.51.3
triton==3.3.0
webdataset==0.2.111